# src/common/api.yml
openapi: 3.0.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
info:
  version: 0.1.0
  title: streamloots
  description: Streamloots Card API
...
consumes:
  - application/json
produces:
  - application/json
...
paths:
...
  /api/v1/cards/{id}:
    get:
      tags:
        - Cards
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the card to retrieve
          type: string
      responses:
        200:
          description: Return the card with the specified id
        404:
          description: Example not found
        401:
          description: Unauthorized          
        403:
          description: Forbidden            
...